calmmage/experiments/seasonal/2024_03_mar/sample-bot/pyproject.toml
[tool.poetry]
name = "sample-bot-lib-bot"
version = "0.0.1" #
description = "Sample bot using bot-lib"
authors = ["Reliable Magician <petr.b.lavrov@gmail.com>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
calmlib = { git = "https://github.com/calmmage/calmlib.git", branch = "main"}


#typer = "*"
#toml = "*"
#pytz = "*"
tqdm = "*"
pyperclip = "*"
#mongoengine = "*"
loguru = "*"
# defaultenv = { git = "https://github.com/bobuk/defaultenv.git", branch ="master", optional = true }
# bmmb = { git = "https://github.com/calmmage/bot_me_maybe.git", branch ="main", optional = true }
#gpt_kit = { git = "https://github.com/calmmage/gpt_kit.git", branch = "main", optional = true }
bot_lib = { git = "https://github.com/calmmage/bot-lib.git", branch = "main" }
#PyGithub = "*"
python-dotenv = "*"
#gitpython = "*"

openai = ">=1.0.0"
langchain = "*"
langchain-anthropic = "*"
langchain-openai = "*"

llama-index = "*"
pydantic-settings = "*"

[tool.poetry.group.dev.dependencies]
jupyter = ">=1.0.0"
pytest = "*"
black = ">=20.8b1"
#pandas = ">=1.0.0"
#pytest = "*"
#pytest-mongodb = "*"
#easydict = ">=1.10"

[tool.poetry.extras]
#full = ["dropbox", "protobuf", "pandas", "seaborn", "defaultenv", "bmmb",
#    "gpt_kit", "httpx", "pydantic", "bokeh", "click", "fastapi", "nltk"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.urls]
#"Home" = "https://github.com/calmmage/calmlib"
